Reorganized and added support for async requests
