+++ Sources/Endpoints/Extensions/URLSession+Endpoints.swift
@@ -31,41 +31,6 @@ extension Endpoint {
     public typealias TaskError = EndpointTaskError<ErrorResponse>
 }
 
-extension Definition {
-    func response(data: Data?, response: URLResponse?, error: Error?) -> Result<Data, T.TaskError> {
-        if let error = error {
-            guard (error as NSError).code != URLError.Code.notConnectedToInternet.rawValue else {
-                return .failure(.internetConnectionOffline)
-            }
-
-            return .failure(.urlLoadError(error))
-        }
-
-        // if we don't have an `error`, we must have an `HTTPURLResponse`
-        guard let httpResponse = response as? HTTPURLResponse else {
-            return .failure(.urlLoadError(URLError(.badServerResponse)))
-        }
-
-        if httpResponse.statusCode == 204 {
-            // handle empty response
-            return .success(Data())
-        } else if (200..<300).contains(httpResponse.statusCode), let data = data {
-            return .success(data)
-        } else if let data = data {
-            let decoded: T.ErrorResponse
-            do {
-                decoded = try T.errorDecoder.decode(T.ErrorResponse.self, from: data)
-            } catch {
-                return .failure(.errorResponseParseError(httpResponse: httpResponse, data: data, error: error))
-            }
-
-            return .failure(.errorResponse(httpResponse: httpResponse, response: decoded))
-        } else {
-            return .failure(.unexpectedResponse(httpResponse: httpResponse))
-        }
-    }
-}
-
 extension URLSession {
 
     /// Creates a session data task using the Definition associated with the passed in request on the passed in environment.
